write to uint8 image in cpp

+++ gym_interf/envs/cpp/include/interflib.h
@@ -1,6 +1,7 @@
 #pragma onces
 
 #include "defines.h"
+#include <cstdint>
 
 extern "C" {
 	EXPORT void calc_image(
@@ -8,6 +9,6 @@ extern "C" {
 		const double*  waveVector1, const double*  center1, double radius1,
         const double*  waveVector2, const double*  center2, double radius2,
         int nFrames, double lambda, double omega, bool hasInterference,
-        int nThreads, double* image);
+        int nThreads, uint8_t* image);
 }
